renamed folder to homepage and homepage return in derived page type
