+++ ws-nextjs-app/pages/[service]/homepage/handleHomepageRoute.ts
